refactor(ethercat): move discovery types to middleware/shared/ports
Mirror of openplc-web's surface refactor so the cross-repo surface
stays byte-identical. Same change: src/types/ethercat/ -> src/middleware/
shared/ports/ethercat-types.ts. Editor-only consumers in main/modules/
ipc/ also updated to the new path.
