+++ src/frontend/components/_features/[workspace]/editor/device/ethercat/components/diagnostics-tab.tsx
@@ -1,7 +1,7 @@
 import { ArrowIcon } from '@root/frontend/assets/icons/interface/Arrow'
 import { cn } from '@root/frontend/utils/cn'
 import type { ScannedDeviceMatch } from '@root/middleware/shared/ports/esi-types'
-import type { EtherCATDevice, NetworkInterface } from '@root/types/ethercat'
+import type { EtherCATDevice, NetworkInterface } from '@root/middleware/shared/ports/ethercat-types'
 
 import { DiscoveredDeviceTable } from './discovered-device-table'
 import { InterfaceSelector } from './interface-selector'
