+++ src/frontend/components/_features/[workspace]/editor/device/ethercat/components/global-settings-tab.tsx
@@ -3,7 +3,7 @@ import { ArrowIcon } from '@root/frontend/assets/icons/interface/Arrow'
 import { InputWithRef } from '@root/frontend/components/_atoms/input'
 import { Select, SelectContent, SelectItem, SelectTrigger } from '@root/frontend/components/_atoms/select'
 import { cn } from '@root/frontend/utils/cn'
-import type { NetworkInterface } from '@root/types/ethercat'
+import type { NetworkInterface } from '@root/middleware/shared/ports/ethercat-types'
 import { useEffect, useState } from 'react'
 
 type GlobalSettingsTabProps = {
