+++ src/middleware/shared/ports/ethercat-types.ts
